requirements.txt
streamlit
PyPDF2
python-dotenv
groq